Refactors Track endpoint and updates docs.
