+++ HeapService.slnx
@@ -1,6 +1,5 @@
 <Solution>
   <Folder Name="/src/">
-    <Project Path="src/Models/Models.csproj" />
     <Project Path="src/Service/Service.csproj" />
   </Folder>
   <Folder Name="/tests/">
