--- src/Models/Class1.cs
@@ -1,6 +1,0 @@
-﻿namespace Models;
-
-public class Class1{
-                   
-                   }
-
