+++ src/Service/Service.json
@@ -5,26 +5,6 @@
     "version": "1.0.0"
   },
   "paths": {
-    "/placeholder": {
-      "get": {
-        "tags": [
-          "test"
-        ],
-        "description": "placeholder",
-        "responses": {
-          "200": {
-            "description": "OK",
-            "content": {
-              "application/json": {
-                "schema": {
-                  "$ref": "#/components/schemas/PayloadOfstring"
-                }
-              }
-            }
-          }
-        }
-      }
-    },
     "/track": {
       "post": {
         "tags": [
@@ -50,21 +30,6 @@
   },
   "components": {
     "schemas": {
-      "PayloadOfstring": {
-        "type": "object",
-        "properties": {
-          "result": {
-            "type": "string",
-            "default": null,
-            "nullable": true
-          },
-          "error": {
-            "type": "string",
-            "default": null,
-            "nullable": true
-          }
-        }
-      },
       "TrackEventParameters": {
         "required": [
           "emailAddress",
@@ -96,9 +61,6 @@
     }
   },
   "tags": [
-    {
-      "name": "test"
-    },
     {
       "name": "heap"
     }
