--- src/Service/docs/Innago.Shared.HeapService.md
@@ -1,9 +1,0 @@
-# Innago.Shared.HeapService assembly
-
-## Innago.Shared.HeapService namespace
-
-| public type | description |
-| --- | --- |
-| record [Payload&lt;T&gt;](./Innago.Shared.HeapService/Payload-1.md) |  |
-
-<!-- DO NOT EDIT: generated by xmldocmd for Innago.Shared.HeapService.dll -->
