--- src/Service/docs/Innago.Shared.HeapService/Payload-1.md
@@ -1,20 +1,0 @@
-# Payload&lt;T&gt; record
-
-```csharp
-public record Payload<T>
-```
-
-## Public Members
-
-| name | description |
-| --- | --- |
-| [Payload](Payload-1/Payload.md)(…) |  |
-| [Error](Payload-1/Error.md) { get; set; } |  |
-| [Result](Payload-1/Result.md) { get; set; } |  |
-
-## See Also
-
-* namespace [Innago.Shared.HeapService](../Innago.Shared.HeapService.md)
-* [Payload.cs](.././Payload.cs)
-
-<!-- DO NOT EDIT: generated by xmldocmd for Innago.Shared.HeapService.dll -->
