--- src/Service/docs/Innago.Shared.HeapService/Payload-1/Error.md
@@ -1,12 +1,0 @@
-# Payload&lt;T&gt;.Error property
-
-```csharp
-public string? Error { get; set; }
-```
-
-## See Also
-
-* record [Payload&lt;T&gt;](../Payload-1.md)
-* namespace [Innago.Shared.HeapService](../../Innago.Shared.HeapService.md)
-
-<!-- DO NOT EDIT: generated by xmldocmd for Innago.Shared.HeapService.dll -->
