--- src/Service/docs/Innago.Shared.HeapService/Payload-1/Payload.md
@@ -1,12 +1,0 @@
-# Payload&lt;T&gt; constructor
-
-```csharp
-public Payload(T? Result = default, string? Error = null)
-```
-
-## See Also
-
-* record [Payload&lt;T&gt;](../Payload-1.md)
-* namespace [Innago.Shared.HeapService](../../Innago.Shared.HeapService.md)
-
-<!-- DO NOT EDIT: generated by xmldocmd for Innago.Shared.HeapService.dll -->
