--- src/Service/docs/Innago.Shared.HeapService/Payload-1/Result.md
@@ -1,12 +1,0 @@
-# Payload&lt;T&gt;.Result property
-
-```csharp
-public T? Result { get; set; }
-```
-
-## See Also
-
-* record [Payload&lt;T&gt;](../Payload-1.md)
-* namespace [Innago.Shared.HeapService](../../Innago.Shared.HeapService.md)
-
-<!-- DO NOT EDIT: generated by xmldocmd for Innago.Shared.HeapService.dll -->
